requirements.txt
python-dotenv
groq
ollama
faiss-cpu
python-pptx
pillow
pdf2image
uvicorn
fastapi
pydantic